requirements.txt
Flask==2.2.5
Pillow==9.5.0
qrcode==7.3.1
requests==2.31.0
python-dotenv==1.0.0
